Prevent any more new HMRC manuals being published in Production
This CSV file lists the slugs for all of HMRC's manuals and the topics which
they should be tagged to (if the publishing of topics is enabled by a feature
flag). We also use this list to restrict the slugs at which manuals can be
published (this restriction is also controlled by a feature flag).

At the moment there are 6 HMRC manuals published in Production. HMRC have
asked us to withdraw 2 of these, and the remaining 4 are the ones we're
leaving in this CSV. They aren't yet ready to publish the rest of the manuals,
and will need to talk to us before they do that, so we're removing the rest
from this CSV.

As far as we know, the data in the rows being removed by this commit is still
accurate, ie the manuals will eventually be published at these slugs and with
these topics, but the topic slugs should be checked again against their
content ids when we later re-add them to this file. There's a rake task to do
that already: `rake map_manual_topic_slugs_to_content_ids`.